rename members to keys
