+++ locksmith/src/graphql/datasource/index.ts
@@ -4,4 +4,4 @@ export * from './keyHolder'
 export * from './keysToRenew'
 export * from './locksByOwner'
 export * from './lock'
-export * from './members'
+export * from './keysByQuery'
